teacher/teacher_details/requirements.txt
asgiref==3.6.0
Django==4.1.7
djangorestframework==3.14.0
numpy==1.24.2
pandas==1.5.3
Pillow==9.4.0
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
